order no done

+++ product/models.py
@@ -23,7 +23,7 @@ class Product(models.Model):
         blank=True, null=True, upload_to="images/product_profile"
     )
     price = models.DecimalField(max_digits=10, decimal_places=2)
-    average_rating = models.FloatField(null=True)
+    average_rating = models.DecimalField(max_digits=3, decimal_places=2, null=True)
 
     def __str__(self):
         return self.name
